requirements.txt
asgiref==3.11.1
Django==5.2.11
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0
sqlparse==0.5.5
